fix(1059): refactor error description, refs: #1059
