+++ frontend/src/enum/SeverityEnum.ts
@@ -9,3 +9,8 @@ export enum ErrorCode {
     SelfGrantProhibited = 'SelfGrantProhibited',
     Default = 'Default',
 }
+
+export const ErrorDescription = {
+    SelfGrantProhibited: 'Granting admin privilege to self is not allowed.',
+    Default: 'An error has occured.',
+}
