+++ frontend/src/store/Constants.ts
@@ -2,8 +2,3 @@ export const FOREST_CLIENT_INPUT_MAX_LENGTH = 8;
 export const FAM_APPLICATION_NAME = 'FAM';
 export const FAM_APPLICATION_ID = 1;
 export const FAM_ADMIN_ROLE = 'FAM_ADMIN';
-
-export const ErrorDescription = {
-    SelfGrantProhibited: 'Granting admin privilege to self is not allowed.',
-    Default: 'An error has occured.',
-}
